Revise Synonym Find to allow Bracketed Objects
Added a Cross Apply to the Synonym finding portion of the script. Replicates the synonym definition four ways. One without brackets, one with brackets on schema, one with brackets on the object, and one with brackets on schema and object. This is to cope with possibility of compiled procedure definitions that use brackets as part of an object name.